analytics_server/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
flasgger==0.9.7.1
PyJWT==2.8.0
requests==2.31.0

